changed collectives so that there are only two sets of psync and pwork arrays
called primary and secondary
should reduce the memory requirements of the code
code alternates between them on successive collectives, so no additional sync is required
to ensure compliance with the standard
there is a use_primary vaiable which controls this, it is flipped after each collective is called

answers are correct for the small problem

+++ definitions.f90
@@ -63,7 +63,7 @@ MODULE definitions_module
 
    LOGICAL      :: advect_x
 
-   INTEGER      :: error_condition
+   REAL(KIND=8) :: error_condition
 
    INTEGER      :: test_problem
    LOGICAL      :: complete
